+++ tests/test_engine.py
@@ -22,41 +22,33 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
 # OR OTHER DEALINGS IN THE SOFTWARE.
 # --------------------------------------------------------------------------
-from __future__ import print_function
+import pytest
 
-import os
-import shutil
-
-from dsdev_utils.paths import ChDir
-
-HTML_DIR = os.path.join(os.getcwd(), 'site')
-DEST_DIR = os.path.join(os.path.expanduser(u'~'), u'BTSync',
-                        u'code', u'Web', u'Menus')
+from menus import BaseMenu, MenusError
+from menus.engine import check_mro, Engine
 
 
-def main():
-    with ChDir(DEST_DIR):
-        files = os.listdir(os.getcwd())
-        for f in files:
-            if f.startswith(u'.'):
-                continue
-            elif f in [u'Procfile', 'Staticfile', 'hostess.json']:
-                continue
-            elif os.path.isfile(f):
-                os.remove(f)
-            elif os.path.isdir(f):
-                shutil.rmtree(f, ignore_errors=True)
+class TestEngine(object):
 
-    with ChDir(HTML_DIR):
-        files = os.listdir(os.getcwd())
-        for f in files:
-            if f.startswith(u'.'):
-                continue
-            if os.path.isfile(f):
-                shutil.copy(f, os.path.join(DEST_DIR, f))
-            elif os.path.isdir(f):
-                shutil.copytree(f, DEST_DIR + os.sep + f)
+    def test_engine_defaults(self):
+        engine = Engine(example=True)
+        assert engine.main.app_name == 'ACME'
 
-if __name__ == '__main__':
-    main()
-    print(u'Move complete')
+
+class TestMRO(object):
+
+    def test_mro_fail(self):
+        class MyMenu(object): pass
+
+        my_menu = MyMenu()
+
+        with pytest.raises(MenusError):
+            check_mro(my_menu)
+
+    def test_mro(self):
+        class MyMenu(BaseMenu):
+            def __init__(self, *args, **kwargs):
+                super(MyMenu, self).__init__(*args, **kwargs)
+
+        my_menu = MyMenu()
+        assert check_mro(my_menu) is True