+++ docs/changelog.md
@@ -7,11 +7,17 @@
 
   - Py3 compat
   - Ability to provide custom menu name
+  - dsdev-utils
+  - A few tests
+  - pause method on BaseMenu
+    - pause(seconds=5, enter_to_continue=False)
 
 ###Updated
 
   - Subclass enforcement
+  - Re-factored internal api
 
 ###Removed
 
   - Requirement to pass MainMenu option
+  - Duplicate code